Rename kas_wgpu::theme::size_handle → dimensions
