+++ kas-wgpu/src/theme/mod.rs
@@ -5,11 +5,11 @@
 
 //! Themes
 
+mod dimensions;
 mod flat_theme;
 mod shaded_theme;
-mod size_handle;
 
-pub(crate) use size_handle::{Dimensions, DimensionsParams, DimensionsWindow};
+pub(crate) use dimensions::{Dimensions, DimensionsParams, DimensionsWindow};
 
 pub use flat_theme::FlatTheme;
 pub use shaded_theme::ShadedTheme;
